src/main/java/com/grasia/prima/ppi/api/service/impl/EmailServiceImpl.java
package com.grasia.prima.ppi.api.service.impl;

import com.grasia.prima.ppi.api.constant.GlobalMessage;
import com.grasia.prima.ppi.api.dto.request.SendEmailRequest;
import com.grasia.prima.ppi.api.exception.BusinessException;
import com.grasia.prima.ppi.api.service.EmailService;
import jakarta.mail.internet.MimeMessage;
import lombok.AllArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.mail.javamail.JavaMailSender;
import org.springframework.mail.javamail.MimeMessageHelper;
import org.springframework.stereotype.Service;

@AllArgsConstructor
@Service
@Slf4j
public class EmailServiceImpl implements EmailService {

    private final JavaMailSender mailSender;
    private static final String FROM_EMAIL = "burnedeadapp@gmail.com";

    @Override
    public void sendEmailHtmlContent(SendEmailRequest request) {
        try {
            MimeMessage message = mailSender.createMimeMessage();
            MimeMessageHelper helper = new MimeMessageHelper(message, true, "UTF-8");
            helper.setFrom(FROM_EMAIL);
            helper.setTo(request.getTo());
            helper.setSubject(request.getSubject());
            helper.setText(request.getBody(), true);

            mailSender.send(message);
            log.info("Mail send success to : {}", request.getTo());
        } catch (Exception e) {
            log.error(e.getMessage());
            throw new BusinessException(GlobalMessage.INTERNAL_SERVER_ERROR);
        }
    }
}
